requirements.txt
mlx-whisper
pyannote.audio
pytest
mutagen
flask-cors
